auth-service/requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
flask-jwt-extended==4.5.3
bcrypt==4.1.2
requests==2.31.0 